feat: ImageRelation 엔티티 연관관계 추가
ImageRelation엔티티에서 Image엔티티로 ManyToOne관계 추가
ImageEntityType Enum클래스를 추가하여 Mysql에서 자동으로 Enum 테이블 생성하도록 변경

Realated to: #7
