+++ src/main/java/com/ktc/togetherPet/model/entity/ImageRelation/ImageEntityType.java
@@ -1,0 +1,7 @@
+package com.ktc.togetherPet.model.entity.ImageRelation;
+
+public enum ImageEntityType {
+    PET,
+    MISSING,
+    REPORT
+}
